Add new workspaces migrations and code refactoring
