+++ src/apps/applets/crud/applets.py
@@ -9,13 +9,14 @@ from sqlalchemy.orm import Query
 from sqlalchemy.sql.functions import count
 
 from apps.applets import errors
-from apps.applets.db.schemas import AppletSchema, UserAppletAccessSchema
+from apps.applets.db.schemas import AppletSchema
 from apps.applets.domain import Role
 from apps.shared.filtering import FilterField, Filtering
 from apps.shared.ordering import Ordering
 from apps.shared.paging import paging
 from apps.shared.query_params import QueryParams
 from apps.shared.searching import Searching
+from apps.workspaces.db.schemas import UserAppletAccessSchema
 from infrastructure.database.crud import BaseCRUD
 
 __all__ = ["AppletsCRUD"]
