+++ src/apps/workspaces/db/schemas/__init__.py
@@ -1,0 +1,2 @@
+from apps.workspaces.db.schemas.user_applet_access import *  # noqa: F401, F403
+from apps.workspaces.db.schemas.user_workspace import *  # noqa: F401, F403
