+++ src/config/__init__.py
@@ -78,5 +78,6 @@ settings = Settings(
         "folders",
         "answers",
         "invitations",
+        "workspaces",
     ],
 )
